Changes in 1.7.0.1 - 5/14/2016: Fixed issue where a bye could cause a crash on pairing.
Changes in 1.7 - 5/13/2016:
Fixed issue where an odd number of tables could fail to print the last odd table when using the Print Tables functionality.
Fixed an issue where leaving the meta blank would cause the last screen of the event to crash.
Added a warning when Regenerating Pairings that all tables will be cleared and regenerated from scratch
Added Name to the game reporting interface
Added a tool tip to the pairings screen to explain the color coding
Moved the next/previous buttons in and down a little bit to avoid interference with Google Hangouts